Updated the slit model and moved it to examples. It now works with all the new updates. I also fixed a bug in beam_line and now have Virac always have the default arguments even when the starting script doesn't specify any.
